requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
openai==1.40.0
pandas==2.1.4
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0